Revises integer types; audits type safety.

+++ cmd/steg/main.go
@@ -120,7 +120,7 @@ func init() {
 		log.Fatalf("atom size must be 1, 2, or 3")
 	}
 
-	state.ctx = steg.NewCtx(*atomSize)
+	state.ctx = steg.NewCtx(uint8(*atomSize))
 	state.carrier, state.carrierSize = getCarrier(*carrier)
 	state.input, state.inputSize = getInput(*input)
 	state.box = *box
